Revert "Merge pull request #1 from teamBunggae/feat/sign-up"
This reverts commit 906dda47b872073927905b0cb6de9544412e17c7.
