+++ .gitignore
@@ -5,8 +5,6 @@ build/
 !**/src/main/**/build/
 !**/src/test/**/build/
 
-application*.yml
-
 ### STS ###
 .apt_generated
 .classpath
