--- .run/RendezvousRestApplication.run.xml
@@ -1,23 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="RendezvousRestApplication" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot" nameIsGenerated="true">
-    <module name="rendezvous.rendezvous-rest.main" />
-    <extension name="coverage">
-      <pattern>
-        <option name="PATTERN" value="com.bunggae.rendezvous.*" />
-        <option name="ENABLED" value="true" />
-      </pattern>
-    </extension>
-    <extension name="software.aws.toolkits.jetbrains.core.execution.JavaAwsConnectionExtension">
-      <option name="credential" />
-      <option name="region" />
-      <option name="useCurrentConnection" value="false" />
-    </extension>
-    <option name="SPRING_BOOT_MAIN_CLASS" value="com.bunggae.rendezvous.RendezvousRestApplication" />
-    <option name="ACTIVE_PROFILES" value="local" />
-    <option name="ALTERNATIVE_JRE_PATH" />
-    <option name="SHORTEN_COMMAND_LINE" value="NONE" />
-    <method v="2">
-      <option name="Make" enabled="true" />
-    </method>
-  </configuration>
-</component>
