+++ build.gradle.kts
@@ -34,20 +34,12 @@ subprojects {
         mavenCentral()
     }
 
-    group = "com.bunggae.rendezvous"
-    version = "0.0.1"
+	group = "com.bunggae.join"
+	version = "0.0.1"
 
     dependencies {
         implementation("org.jetbrains.kotlin:kotlin-reflect")
         implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
-
-        implementation("io.github.microutils:kotlin-logging:_")
-        implementation("com.fasterxml.jackson.module:jackson-module-kotlin:_")
-
-        testImplementation("io.kotest:kotest-runner-junit5:_")
-        testImplementation("io.kotest:kotest-assertions-core:_")
-        testImplementation("io.kotest:kotest-property:_")
-        testImplementation("io.kotest:kotest-framework-datatest:_")
     }
 }
