+++ join-core/src/main/kotlin/com/bunggae/join/user/domain/User.kt
@@ -1,0 +1,4 @@
+package com.bunggae.join.user.domain
+
+class User {
+}
