+++ join-data/build.gradle.kts
@@ -1,0 +1,9 @@
+dependencies {
+    implementation(project(":join-core"))
+}
+
+val jar: Jar by tasks
+val bootJar: org.springframework.boot.gradle.tasks.bundling.BootJar by tasks
+
+bootJar.enabled = false
+jar.enabled = true
