+++ join-rest/build.gradle.kts
@@ -1,0 +1,9 @@
+dependencies {
+    implementation(project(":join-core"))
+    implementation(project(":join-data"))
+
+    implementation("org.springframework.boot:spring-boot-starter")
+    implementation("org.springframework.boot:spring-boot-starter-web")
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+
+}
