+++ join-rest/src/main/kotlin/com/bunggae/join/JoinRestApplication.kt
@@ -1,0 +1,11 @@
+package com.bunggae.join
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class JoinRestApplication
+
+fun main(vararg args: String) {
+     runApplication<JoinRestApplication>(*args)
+}
