--- rendezvous-core/build.gradle.kts
@@ -1,9 +1,0 @@
-dependencies {
-    implementation("javax.inject:javax.inject:_")
-}
-
-val jar: Jar by tasks
-val bootJar: org.springframework.boot.gradle.tasks.bundling.BootJar by tasks
-
-bootJar.enabled = false
-jar.enabled = true
