--- rendezvous-core/src/main/kotlin/com/bunggae/rendezvous/user/application/infra/MailService.kt
@@ -1,5 +1,0 @@
-package com.bunggae.rendezvous.user.application.infra
-
-interface MailService {
-    fun sendVerificationEmail(targetEmail: String, code: String)
-}
