--- rendezvous-core/src/main/kotlin/com/bunggae/rendezvous/user/application/usecase/CheckAlreadyExistingEmailUseCase.kt
@@ -1,16 +1,0 @@
-package com.bunggae.rendezvous.user.application.usecase
-
-import com.bunggae.rendezvous.user.application.aggregate.UserAggregate
-import javax.inject.Named
-
-@Named
-class CheckAlreadyExistingEmailUseCase(
-    private val userAggregate: UserAggregate,
-) {
-    data class Query(
-        val email: String,
-    )
-
-    fun execute(query: Query): Boolean =
-        userAggregate.findByEmailOrNull(query.email) != null
-}
