--- rendezvous-data/src/main/kotlin/com/bunggae/rendezvous/base/JpaBaseEntity.kt
@@ -1,10 +1,0 @@
-package com.bunggae.rendezvous.w
-
-import java.time.LocalDateTime
-import javax.persistence.MappedSuperclass
-
-@MappedSuperclass
-open class JpaBaseEntity(
-    var createdAt: LocalDateTime = LocalDateTime.now(),
-    var updatedAt: LocalDateTime = LocalDateTime.now()
-)
