--- rendezvous-data/src/main/kotlin/com/bunggae/rendezvous/base/SoftDeletable.kt
@@ -1,7 +1,0 @@
-package com.bunggae.rendezvous.base
-
-import java.time.LocalDateTime
-
-interface SoftDeletable {
-    var deletedAt: LocalDateTime?
-}
