--- rendezvous-data/src/main/kotlin/com/bunggae/rendezvous/user/adapter/repository/UserJpaRepository.kt
@@ -1,9 +1,0 @@
-package com.bunggae.rendezvous.user.adapter.repository
-
-import com.bunggae.rendezvous.user.record.UserJpaRecord
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserJpaRepository: JpaRepository<UserJpaRecord, Long> {
-    fun findByEmail(email: String): UserJpaRecord?
-    fun findByServiceId(serviceId: String): UserJpaRecord?
-}
