--- rendezvous-infra/build.gradle.kts
@@ -1,15 +1,0 @@
-dependencies {
-    implementation(project(":rendezvous-core"))
-
-    implementation("javax.inject:javax.inject:_")
-
-    // mail
-    implementation("com.sun.mail:jakarta.mail:_")
-    implementation("com.sun.activation:jakarta.activation:_")
-}
-
-val jar: Jar by tasks
-val bootJar: org.springframework.boot.gradle.tasks.bundling.BootJar by tasks
-
-bootJar.enabled = false
-jar.enabled = true
