--- rendezvous-rest/src/main/kotlin/com/bunggae/rendezvous/config/InfraEnvConfig.kt
@@ -1,21 +1,0 @@
-package com.bunggae.rendezvous.config
-
-import com.bunggae.rendezvous.InfraEnvHolder
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.core.env.Environment
-
-@Configuration
-class InfraEnvConfig(
-    private val emailCredentialProperties: EmailCredentialProperties,
-) {
-    @Bean
-    fun infraEnvHolder(
-        env: Environment,
-    ): InfraEnvHolder {
-        val infraEnvHolder = InfraEnvHolder(env.activeProfiles)
-        val (address, password) = emailCredentialProperties
-        infraEnvHolder.setEmailConfig(address, password)
-        return infraEnvHolder
-    }
-}
