--- rendezvous-rest/src/main/kotlin/com/bunggae/rendezvous/user/dto/CheckExistingServiceIdReqDto.kt
@@ -1,5 +1,0 @@
-package com.bunggae.rendezvous.user.dto
-
-data class CheckExistingServiceIdReqDto(
-    val serviceId: String
-)
