--- rendezvous-rest/src/main/kotlin/com/bunggae/rendezvous/user/dto/UserSignUpReqDto.kt
@@ -1,10 +1,0 @@
-package com.bunggae.rendezvous.user.dto
-
-data class UserSignUpReqDto(
-    val email: String,
-    val password: String,
-    val userName: String,
-    val serviceId: String,
-    val profileImgUrl: String,
-)
-
