+++ settings.gradle.kts
@@ -1,10 +1,4 @@
-rootProject.name = "rendezvous"
-include("rendezvous-core")
-include("rendezvous-rest")
-include("rendezvous-data")
-include("rendezvous-infra")
-
-plugins {
-    // See https://jmfayard.github.io/refreshVersions
-    id("de.fayard.refreshVersions") version "0.40.1"
-}
+rootProject.name = "join"
+include("join-core")
+include("join-rest")
+include("join-data")
