--- versions.properties
@@ -1,30 +1,0 @@
-#### Dependencies and Plugin versions with their available updates.
-#### Generated by `./gradlew refreshVersions` version 0.40.1
-####
-#### Don't manually edit or split the comments that start with four hashtags (####),
-#### they will be overwritten by refreshVersions.
-####
-#### suppress inspection "SpellCheckingInspection" for whole file
-#### suppress inspection "UnusedProperty" for whole file
-
-version.com.fasterxml.jackson.module..jackson-module-kotlin=2.13.2
-
-version.com.sun.activation..jakarta.activation=2.0.1
-
-version.com.sun.mail..jakarta.mail=2.0.1
-
-version.io.github.microutils..kotlin-logging=2.1.21
-
-version.jakarta.mail..jakarta.mail-api=2.1.0
-
-version.javax.inject..javax.inject=1
-
-version.kotest=5.2.3
-
-version.mysql..mysql-connector-java=8.0.29
-
-version.org.springframework.boot..spring-boot-configuration-processor=2.6.7
-
-version.org.springframework.boot..spring-boot-starter-data-jpa=2.6.7
-
-version.org.springframework.boot..spring-boot-starter-mail=2.6.7
